analytics/producer/copilot/bitcoin-feed/requirements.txt
boto3
pyzmq
awswrangler
requests